--- opentelekomcloud/acceptance/iam/import_opentelekomcloud_identity_user_v3_test.go
@@ -1,39 +1,0 @@
-package acceptance
-
-import (
-	"fmt"
-	"testing"
-
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/acctest"
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
-
-	"github.com/opentelekomcloud/terraform-provider-opentelekomcloud/opentelekomcloud/acceptance/common"
-)
-
-func TestAccIdentityV3User_importBasic(t *testing.T) {
-	resourceName := "opentelekomcloud_identity_user_v3.user_1"
-	var userName = fmt.Sprintf("ACCPTTEST-%s", acctest.RandString(5))
-
-	resource.Test(t, resource.TestCase{
-		PreCheck: func() {
-			common.TestAccPreCheck(t)
-			common.TestAccPreCheckAdminOnly(t)
-		},
-		ProviderFactories: common.TestAccProviderFactories,
-		CheckDestroy:      testAccCheckIdentityV3UserDestroy,
-		Steps: []resource.TestStep{
-			{
-				Config: testAccIdentityV3UserBasic(userName),
-			},
-
-			{
-				ResourceName:      resourceName,
-				ImportState:       true,
-				ImportStateVerify: true,
-				ImportStateVerifyIgnore: []string{
-					"password",
-				},
-			},
-		},
-	})
-}
